+++ src/main/java/com/tencentcloudapi/as/v20180419/models/ModifyScheduledActionRequest.java
@@ -41,21 +41,21 @@ public class ModifyScheduledActionRequest extends AbstractModel{
     */
     @SerializedName("MaxSize")
     @Expose
-    private Integer MaxSize;
+    private Long MaxSize;
 
     /**
     * 当定时任务触发时，设置的伸缩组最小实例数。
     */
     @SerializedName("MinSize")
     @Expose
-    private Integer MinSize;
+    private Long MinSize;
 
     /**
     * 当定时任务触发时，设置的伸缩组期望实例数。
     */
     @SerializedName("DesiredCapacity")
     @Expose
-    private Integer DesiredCapacity;
+    private Long DesiredCapacity;
 
     /**
     * 定时任务的首次触发时间，取值为`北京时间`（UTC+8），按照`ISO8601`标准，格式：`YYYY-MM-DDThh:mm:ss+08:00`。
@@ -114,7 +114,7 @@ public class ModifyScheduledActionRequest extends AbstractModel{
      * Get 当定时任务触发时，设置的伸缩组最大实例数。 
      * @return MaxSize 当定时任务触发时，设置的伸缩组最大实例数。
      */
-    public Integer getMaxSize() {
+    public Long getMaxSize() {
         return this.MaxSize;
     }
 
@@ -122,7 +122,7 @@ public class ModifyScheduledActionRequest extends AbstractModel{
      * Set 当定时任务触发时，设置的伸缩组最大实例数。
      * @param MaxSize 当定时任务触发时，设置的伸缩组最大实例数。
      */
-    public void setMaxSize(Integer MaxSize) {
+    public void setMaxSize(Long MaxSize) {
         this.MaxSize = MaxSize;
     }
 
@@ -130,7 +130,7 @@ public class ModifyScheduledActionRequest extends AbstractModel{
      * Get 当定时任务触发时，设置的伸缩组最小实例数。 
      * @return MinSize 当定时任务触发时，设置的伸缩组最小实例数。
      */
-    public Integer getMinSize() {
+    public Long getMinSize() {
         return this.MinSize;
     }
 
@@ -138,7 +138,7 @@ public class ModifyScheduledActionRequest extends AbstractModel{
      * Set 当定时任务触发时，设置的伸缩组最小实例数。
      * @param MinSize 当定时任务触发时，设置的伸缩组最小实例数。
      */
-    public void setMinSize(Integer MinSize) {
+    public void setMinSize(Long MinSize) {
         this.MinSize = MinSize;
     }
 
@@ -146,7 +146,7 @@ public class ModifyScheduledActionRequest extends AbstractModel{
      * Get 当定时任务触发时，设置的伸缩组期望实例数。 
      * @return DesiredCapacity 当定时任务触发时，设置的伸缩组期望实例数。
      */
-    public Integer getDesiredCapacity() {
+    public Long getDesiredCapacity() {
         return this.DesiredCapacity;
     }
 
@@ -154,7 +154,7 @@ public class ModifyScheduledActionRequest extends AbstractModel{
      * Set 当定时任务触发时，设置的伸缩组期望实例数。
      * @param DesiredCapacity 当定时任务触发时，设置的伸缩组期望实例数。
      */
-    public void setDesiredCapacity(Integer DesiredCapacity) {
+    public void setDesiredCapacity(Long DesiredCapacity) {
         this.DesiredCapacity = DesiredCapacity;
     }
